requirements.txt
Django==1.7.4
Pillow==2.7.0
argparse==1.2.1
django-simple-captcha==0.4.4
six==1.9.0
wsgiref==0.1.2
